chore(common): Add PlatformExpressHandler abstraction
