+++ packages/common/src/platform-express/index.ts
@@ -15,5 +15,6 @@ export * from "./decorators/ExpressRouter";
 export * from "./services/ServeStaticService";
 export * from "./services/PlatformExpressApplication";
 export * from "./services/PlatformExpressRouter";
+export * from "./services/PlatformExpressHandler";
 // UTILS
 export * from "./utils";
