+++ packages/common/src/platform/services/PlatformHandler.spec.ts
@@ -156,8 +156,6 @@ describe("PlatformHandler", () => {
         // THEN
         expect(result).to.eq(undefined);
         handler.length.should.eq(3);
-        Test.prototype.get.should.have.been.calledWithExactly("testValue");
-        request.ctx.data.should.eq("testValue");
       })
     );
     it(
@@ -294,98 +292,6 @@ describe("PlatformHandler", () => {
 
         // THEN
         expect(value).to.deep.eq(context.err);
-      })
-    );
-    it(
-      "should return BODY",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.BODY);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.request.body);
-      })
-    );
-    it(
-      "should return PATH",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.PATH);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.request.params);
-      })
-    );
-    it(
-      "should return QUERY",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.QUERY);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.request.query);
-      })
-    );
-    it(
-      "should return HEADER",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.HEADER);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq({"content-type": "application/json"});
-      })
-    );
-    it(
-      "should return COOKIES",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.COOKIES);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.request.cookies);
-      })
-    );
-    it(
-      "should return SESSION",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.SESSION);
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.request.session);
-      })
-    );
-    it(
-      "should return LOCALS",
-      PlatformTest.inject([InjectorService, PlatformHandler], async (injector: InjectorService, platformHandler: PlatformHandler) => {
-        // GIVEN
-        const {param, context} = build(injector, ParamTypes.LOCALS);
-        context.err = new Error();
-
-        // WHEN
-        const value = platformHandler.getParam(param, context);
-
-        // THEN
-        expect(value).to.deep.eq(context.response.locals);
       })
     );
     it(
